feat(init.lua): add colorcolumn setting and specify Python 3 host prog
This commit includes two enhancements to the init.lua configuration file for Vim. First, it introduces a 'colorcolumn' setting to highlight the 79th column, which is a common practice for indicating the maximum line length for better code readability. Second, it specifies the Python 3 host program path for Neovim's Python integration by setting the 'python3_host_prog' variable. This will assist users in setting up a consistent Python environment for plugins requiring Python 3 support.

In addition, this commit removes 'dart' from the 'ensure_installed' list within the nvim-treesitter configuration, streamlining the setup for the languages most commonly used in this configuration.