refactor: split complex `init.lua` into multiple modules
This commit refactors the neovim `init.lua` configuration file by separating
its various configurations into multiple modular files. The different areas
of configuration addressed include options, keybindings, LSP setup,
Telescope setup, Treesitter setup, and completion setup with nvim-cmp.
Each of these configurations have been extracted from `init.lua` and
placed into their respective `lua/*.lua` files.

As a result of this refactoring, the `init.lua` file becomes more
maintainable and easier to navigate. Contributions to different parts of
the configuration can now be done in a more isolated fashion, reducing
the risk of conflicts during collaboration and making the overall
codebase more structured.

This modular approach also follows the best practices in software
engineering where separation of concerns is encouraged. It allows users
of the configuration to more clearly understand the purpose of each
section, and to focus on one area at a time, thereby enhancing the
quality of the configuration and the ease of making changes.

Whenever a configuration snippet is related to a particular functionality,
for instance LSP or completion, the snippet is moved to its corresponding
file: `lsp-setup.lua` for LSP configurations or `cmp-setup.lua` for
completion configurations.

Summary of changes:
- Extracted options into `lua/options.lua`
- Extracted keybindings into `lua/keymaps.lua`
- Moved LSP configurations to `lua/lsp-setup.lua`
- Moved Telescope-specific settings to `lua/telescope-setup.lua`
- Treesitter configuration is now in `lua/treesitter-setup.lua`
- Nvim-cmp setup, including snippet and kind icons, is moved to
  `lua/cmp-setup.lua`

With clear-cut boundaries between different aspects of the configuration,
the `.lua` files can be independently maintained, tested, and reused.