feat(treesitter-setup): disable indent and textobjects for Dart due to delays
This commit updates the treesitter setup in `treesitter-setup.lua` to address
performance issues when editing Dart files. A significant delay was observed
when creating new lines in Dart files, as reported in multiple issues [1][2][3].
To mitigate this, the indent and textobjects configurations for Dart have been
disabled. This should prevent the editor from hanging and improve the overall
user experience when working with Dart code.

The commit also includes a small change where 'rust' was removed from the list
of languages in `ensure_installed` to match the project's current requirements.
The line was refactored for consistency.

References:
- [1] https://github.com/UserNobody14/tree-sitter-dart/issues/48
- [2] https://github.com/UserNobody14/tree-sitter-dart/issues/46
- [3] https://github.com/nvim-treesitter/nvim-treesitter/issues/4945